+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.14
 ### is.codion.common.core
 - PropertyStore, property access now based on predicates instead of key prefixes.
+### is.codion.swing.framework.ui
+- EntityTablePanel bug fixed, Config.SHOW_REFRESH_PROGRESS_BAR now used for default config value.
 
 ## 0.18.13
 ### is.codion